requirements.txt
validators
beautifulsoup4
lxml
requests
pysocks
pymongo
tldextract
python-magic
chrome-cookiejar
pandas
